+++ src/transforms/ember-getWithDefault-to-lodash-get/tests/fixtures/lodash.get.input.js
@@ -1,0 +1,6 @@
+import _get from 'lodash-es/get';
+import {getWithDefault} from '@ember/object';
+
+const a = _get(foo, 'bar');
+const b = getWithDefault(foo, 'bar', 'baz');
+const c = this.getWithDefault(foo, 'bar');