+++ src/transforms/ember-getWithDefault-to-lodash-get/tests/fixtures/lodash.get.output.js
@@ -1,0 +1,5 @@
+import _get from 'lodash-es/get';
+
+const a = _get(foo, 'bar');
+const b = _get(foo, 'bar', 'baz');
+const c = _get(this, foo, 'bar');