+++ src/transforms/ember-getWithDefault-to-lodash-get/tests/fixtures/lodash.import.output.js
@@ -1,0 +1,6 @@
+import getWithDefault from 'lodash/get';
+import _ from 'lodash';
+
+const a = _.set(foo, 'bar', baz);
+const b = getWithDefault(foo, 'bar', baz);
+const c = getWithDefault(this, 'bar', baz);