+++ src/transforms/ember-getWithDefault-to-lodash-get/tests/fixtures/transform.input.js
@@ -1,0 +1,9 @@
+import {computed, getWithDefault} from '@ember/object';
+
+export default Component.extend({
+  test: computed('foo', function () {
+    const bar = getWithDefault(foo, 'bar', '');
+    const baz = this.getWithDefault(qux, '');
+    return [bar, baz]
+  }).readOnly()
+})