+++ src/transforms/ember-getWithDefault-to-lodash-get/tests/fixtures/transform.output.ts
@@ -1,0 +1,4 @@
+// @ts-nocheck
+import getWithDefault from 'lodash/get';
+import a from 'b';
+getWithDefault(this, foo, 'bar');