+++ src/transforms/remove-global-ember/tests/fixtures/has-ember-imports.output.ts
@@ -1,0 +1,12 @@
+import Ember, {other} from 'ember';
+import {something} from 'ember'
+import {Component} from 'frost-core-components'
+
+const {testing} = Ember
+
+export default Component.extend({
+  // == Functions =============================================================
+  getTesting (): boolean {
+    return testing && other && something
+  }
+})