+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ciena-org/ember-codemods
 
+## 0.3.2
+
+### Patch Changes
+
+- c8dc45f: Add ember-getWithDefault-to-lodash-get codemode for ember-metal.getWithDefault deprecation
+- b836e2a: feat: Add has-block transform to address has-block-and-has-block-params deprecation
+
+## 0.3.1
+
+### Patch Changes
+
+- 3920191: feat: Add remove-global-ember transform
+
 ## 0.3.0
 
 ### Minor Changes